requirements.txt
aiohttp>=3.9.4,<4.0.0
aiofiles>=23.2.1,<24.0.0
ffmpeg-python>=0.2.0,<1.0.0
gTTS>=2.5.1,<3.0.0
dnspython>=2.1.0,<3.0.0
future>=1.0.0,<2.0.0
googletrans>=4.0.0-rc1,<5.0.0
motor>=3.6,<4.0.0
pillow>=10.0.0,<11.0.0
psutil>=5.9.8,<6.0.0
pykeyboard>=0.1.5,<0.2.0
python-arq>=6.0.7,<7.0.0
requests>=2.31.0,<3.0.0
search-engine-parser>=0.6.8,<0.7.0
speedtest-cli>=2.1.3,<3.0.0
TgCrypto>=1.2.5,<2.0.0
uvloop>=0.19.0,<1.0.0
youtube_dl>=2021.12.17,<2022.0.0
beautifulsoup4>=4.12.2,<5.0.0
python-dotenv>=1.0.1,<2.0.0
feedparser>=6.0.11,<7.0.0
pyromod>=3.1.6,<4.0.0
fuzzysearch>=0.7.3,<0.8.0
img2pdf>=0.5.1,<0.6.0
telegram-telegraph>=2.2.0,<3.0.0
pytube>=15.0.0,<16.0.0
pytz>=2024.1,<2025.0
kurigram>=2.2.12,<3.0.0
pyrogram>=2.0.0,<3.0.0
python-socks>=2.0.0,<3.0.0
pymongo>=4.6.0,<5.0.0
pyrogram-raw-core>=2.0.0,<3.0.0
pyrogram-raw-types>=2.0.0,<3.0.0
pyrogram-raw-methods>=2.0.0,<3.0.0
python-multipart>=0.0.6,<0.1.0
aiohttp-socks>=0.8.3,<0.9.0
